+++ demo/config.js
@@ -1,0 +1,11 @@
+/* eslint-disable import/no-extraneous-dependencies, import/no-unresolved, import/extensions */
+
+import { configure } from '@storybook/angular';
+
+// automatically import all files ending in *.demo.ts
+const req = require.context('.', true, /.demo.ts$/);
+function loadStories() {
+    req.keys().forEach((filename) => req(filename));
+}
+
+configure(loadStories, module);