+++ gulpfile.js
@@ -20,18 +20,19 @@ const styles = `${dist}/styles`;
  * Main tasks
  */
 
-gulp.task('build', sequence('clean', 'prepare-build', ['compile', 'clean-temp', 'styles'], 'bundle', 'minify'));
+gulp.task('build', sequence('clean-dist', 'prepare-build', 'compile', ['clean-temp', 'styles'], 'bundle', 'minify'));
 gulp.task('deploy', sequence('build', 'prepare-deploy', 'publish'));
 gulp.task('demo', ['prepare-demo'], () => {
     gulp.watch('./src/**/*.demo.ts', ['copy-demo'])
     gulp.watch(['./src/**/*.html', './src/**/*.ts', '!./**/*.spec.ts', '!./**/*.demo.ts'], ['compile-demo'])
 });
+gulp.task('clean', ['clean-dist', 'clean-temp', 'clean-demo']);
 
 /**
  * Aux tasks
  */
 
-gulp.task('clean', cb => remove(dist, cb));
+gulp.task('clean-dist', cb => remove(dist, cb));
 gulp.task('clean-temp', cb => remove(temp, cb));
 gulp.task('clean-demo', cb => remove(demo, cb));
 gulp.task('prepare-build', () =>