+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+## Current
+
+### Bug fixes
+
+- Documentation mistakes fixed
+
 ## **[1.0.0]** - 16/01/2018
 
 [1.0.0]: https://github.com/nitsnets/unicorn_components/releases/tag/v1.0.0 "v1.0.0"