+++ CHANGELOG.md
@@ -2,9 +2,10 @@
 
 ## Current
 
-### Bug fixes
 
-- [Storybook] integrated for the demo.
+## **[1.0.1]** - 25/01/2018
+### Bug fixes
+- Storybook integrated for the demo.
 - Documentation mistakes fixed
 
 ## **[1.0.0]** - 16/01/2018
@@ -12,3 +13,4 @@
 [Storybook]: https://github.com/storybooks/storybook
 
 [1.0.0]: https://github.com/nitsnets/unicorn_components/releases/tag/v1.0.0 "v1.0.0"
+[1.0.1]: https://github.com/nitsnets/unicorn_components/releases/tag/v1.0.1 "v1.0.0"